requirements.txt
fastapi-mcp==0.3.4
uvicorn==0.34.1
FastMCP==2.2.0
pandas==2.2.3
joblib==1.5.1
scikit-learn==1.6.1
streamlit==1.45.1
anthropic==0.52.1
fastapi==0.115.12